front: move operationalStudies/components/Scenario content in operationalStudies/views/Scenario
