+++ front/src/applications/operationalStudies/views/Scenario/components/BoardWrapper.tsx
