+++ front/src/applications/operationalStudies/views/Scenario/components/EditedElementContainerContext.tsx
