+++ front/src/applications/operationalStudies/views/Scenario/components/InfraLoadingState.tsx
